requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Pillow==10.0.1
tensorflow==2.15.0
tensorflow_hub==0.15.0
matplotlib==3.8.4
python-telegram-bot==20.3